Fixed the settings being used as dwords instead of bytes

+++ AAUnlimited/Functions/AAPlay/GameState.cpp
@@ -52,7 +52,7 @@ struct GameStateStruct {
 	DWORD m_PCConversationState;		//0 = still speaking, 1 = waiting for answer, 2/3 = answering/end?
 	DWORD m_NPCLineState;				//increments from 0 to whatever
 	int roomNumber[25];					//Current room ID
-	DWORD interrupt;					//Disabled interruptions
+	BYTE interrupt;					//Disabled interruptions
 	bool h_ai;							//Disable or enable h-ai
 	bool h_ai_locked;					//Disable or enable ability to leave h-ai
 	DWORD m_HPosition;					//H position ID
@@ -265,7 +265,7 @@ void Shared::GameState::setVoyeurTarget(ExtClass::NpcData* target)
 	loc_gameState.m_voyeurTarget = target;
 }
 
-void Shared::GameState::setInterrupt(DWORD value)
+void Shared::GameState::setInterrupt(BYTE value)
 {
 	loc_gameState.interrupt = value;
 }
